Move config.js to lib folder
